Recoding the preprocessing architecture so that stem and loop are preprocessed simultaneously.
A LassoBuilder object is handed through all preprocessors. This object superseeds the VarFactory and spits out a Lasso object in the end.

The new TransFormulaLR object is LassoRanker's analogue to the TransFormula. It superseeds the VarCollector.

Every LassoBuilder holds a list of TransFormulaLR for the stem (the stem's connected components) and a list of TransFormulaLR for the loop (the loop's connected components). Note that the code for supporting connected components is very limited at this point, e.g. when extracting a Lasso from LassoBuilder at this point it is assumed that there is only one component.