Deactivate the tray icon
I did not dare to remove the code since someone might still want to have
tray icons.  My suggestion in this case:  leave the shell visible since
making it invisible causes problems on linux.  See comments added to the
ApplicationWorkbenchWindowAdvisor.  Also fixed the TrayIconNotifier to
prevent NPEs in the GUI thread.  Should not harm, but are useless and
just waste resources.