Major refactoring of typesize and offset computation
New class TypeSizeAndOffsetComputer, which does computations that were
formerly done in MemoryHandler and StructHandler.

The new class provides a unified computation for the fixed and non-fixed
typesize settings. The former is now also compatible to the "bitvector
translation" the latter is still not yet useable.

For the fixed typesize setting we now omit typesize/offset constants
and axioms completely. We now also use the computed typesizes/offset
when accessing the heap.