add support for "deprecated predicates" to Predicate Unifier
Assume we want to construct an IPredicate for a given term,
the term is quantifier free and equivalent to the term of an existing
IPredicate p that is not quantifier free.

Before this commit we just returned p.
Now we construct a new IPredicate p' with the quantifier free term and
add p->p' to a map which indicated that p is deprecated.
The old prediate is still in the coverage relation, but should never be
returned while constructing new IPredicates