Abstract Interpretation Interval Division
The result of the interval division of an interval with another interval
that contains zero now leads to a new interval (-infty; infty) as
anything may happen when performing a division by zero. Note that this
is unsound if there is no check for a division by zero introduced into
the original program.