Abstract Interpretation Overapproximative Array Handling
Added overapproximative array handling in nonrelational states for
abstract interpretation. Now, the exception "Proper array handling not
implemented" should not be thrown anymore, as all array accesses are
over-approximated with TOP.

Note that in the class "NonrelationalStatementProcessor" there is the
method "processExpression" which should be modified as soon as proper
array handling is implemented. See the TODO at the respective locations
in the methods.