Abstract Interpretation Precision Improvement for Nonrelational Domains
Implemented handling of functions whose return values are dependent on
the range of the input expressions. This currently only works in
nonrelational domains.

To this end, the nested class ConstOnlyIdentifierTranslator in
Boogie2SMT has been made public, as well as a method was added that
allows to return the mExpression2Term object.