Abstract Interpretation BoogieConst in case of BoogieVar == null
Returning the corresponding BoogieConst if no BoogieVar can be found in
the Boogie2SMTSymbolTable. This happens if, e.g., a program uses
function pointers which are translated to constants that were added as
consts instead of "normal" variables.

Also added a few failsafe asserts which should catch all cases where the
queried variable is not present.