Abstract Interpretation Nonrelational States Fixes
Removed assertion that checks whether a state not marked as bottom does
only contain values that are not bottom, if the state is created. This
led to multiple errors in the logic while creating abstract states.

The removal of the assertion does not have an effect for soundness or
correctness, because the isBottom() method checks the values of the
variables anyways before returning the previously set isBottom
indicator.