Generalization of abstract states
Added generic parameter to abstract states that determines the type of
the used program variables (in our case so far: IBoogieVar, and
IProgramVar).

This is the groundwork to introducing abstract domains that can handle
transformulas. By this change, only the main class of each abstract
domain and the post operator of the abstract domains need to be written
in order to support transformulas. The rest can stay the same.