Fixed a bug in abstract interpretation's nonrelational domain
The bug resulted in the computation of LOGICAND evaluation to be an
over-approximation, which was unnecessary.

This fix, however, broke at least one regression test (callABAB).
Currently, it is unclear whether that fix broke the regression test on a
theoretical level or whether there are some other bugs that only become
apparent now.