Made the nonrelational domain more precise when evaluating assumes
If an assume statement is evaluated, the expected outcome of the assume
is now always "true", since the outcome "false" would mean that the
successor state is bottom. Before, "top" was also possible in some cases
which is not wrong, but an unnecessary over-approximation.

Also added a regression test that uncovers this behavior.