BugFix: matching letters when reading FH automata from files
Bug was that FH-automata that were constructed for some program P, were not sufficient to prove P correct after reading them back from ats files.
Instead of mapping a string-letter to a single corresponding object-letter (LETTER), we now map it to *all* corresponding object-letters.
This increases the number of transitions in the automaton and makes the interpretion of FH automata slower, but avoids this bug.

Signed-off-by: Bat-Chen <batcheni89@gmail.com>