optimization in ICFG builder
If a call would be turned into a summary edge (because the called
procedure does not have an implementation) and there is no non-free
requires (that would require the construction of an error location),
then we can now add the call into a sequential composition.

This will help me to support atomic blocks in the input code. But as a
side-effect there is less work that has to be done by our block encoding
algorithms.