Added printf logging methods (more details below). Improved implementation to use enums instead of logXXX methods internally.
In Ultimate we often use the following pattern.

    if (logger.isDebugEnabled()) {
        // obj's toString() method is expensive -- only call it if really needed
        logger.debug("..." + obj + "...");
        logger.debug(obj);
    }

The new logger does the check for us. We don't have to use a surrounding "if" anymore.
After this change the following code is equivalent to the one above.

    logger.debug("... %s ...", obj);
    logger.debug(obj);

Note that `logger.debug("..." + obj + "...");` would still call obj.toString() even if
the debug log level was disabled.

The new printf style can also be used instead of long string concatenations.
In the past people seemed to be reluctant to write `log.debug(String.format(...))`
and prefered to write long chains like `log.debug("…" + a + "…" + b + "…" + c + "…" + ...)`.