Fixed wrong usage of call summarizer.
Call summarizers can only use the input AFTER THE CALL STATEMENT (that is, a predicate without local variables of the caller).
However, until now we used only the inpute BEFORE THE CALL STATEMENT (that is, a predicate without parameters / local variables of the callee).
The bug caused no trouble so far because our only summarizer ignored the input and always used TOP.