fix CfgBuilder handling of Y-to-V compositions
Since support for Y-to-V was introduced, we were unable to
build CFGs for many examples in pthread-wmm (SV-COMP).
The reason was an exponential number of edges being created
due to eager Y-to-V compositions.

This fix only allows Y-to-V when no other compositions are
possible. We detect atomic block nodes early, so that normal
sequential compositions are possible.