do not abstract unmodified variables that are partially constrained
If a variable that is not modified by a statement appears only
in the inConstraints, or only in the outConstraints, we treat it
as if it appeared in both.

The reason is that we can not abstract it properly without either

(1) breaking soundness wrt to valid Hoare triples, or
(2) making it an assigned variable of the abstracted statement.

For now (in this commit), we choose to preserve (2) [and of course, (1)]
and instead sacrifice the property that the abstracted statement read only
in-constrained variables.