fix locked ManagedScript in VariableAbstraction
CfgSmtToolkit's ManagedScript may be locked by Floyd/Hoare automaton,
in particular if enhancement is not NONE. This is a problem when the
SpecificVariableAbstraction e.g. declares new constants for auxVars:
An exception is thrown.

This bug was hidden by the fact that enabled assertions unlock the
ManagedScript before the exception occurs.

As a solution, we let the SpecificVariableAbstraction already transfer
TransFormulas to the script used for commutativity checking. We still
need to transfer conditions (in case of conditional independence).

If no abstraction is used, both TransFormulas and conditions are transferred
by the SemanticIndependenceRelation as before.