improve SpecificVariableAbstraction::refine
With the previous implementation, caching was severely limited:
As soon as the constraints for any letter was refined, the new
constraint was strictly less than the constrained returned by
previous calls to ::restrict, and the abstractions for all letters
were discarded.

The fixed implementation allows the constraints for other letters
to be refined all the way to the bottom-most abstraction level,
without invalidating the cached abstraction. Only if the constraints
for the cached letter are refined beyond a certain point, the cached
abstraction is invalidated.