Adapt auxvar elimination to new quantifier elimination
It looks like this method expects a quantifier elimination that states
which quantified variables were successfully eliminated. Our new (few
years old) quantifier elimination does not have such a feature any more.
Hence it does not make sense to have a method for eliminating auxvars
that returns variables that could not be eliminated.