remove LETTER type parameter from Marking
A marking is a set of places -- so why is it parameterized in a
type of letters? It is sufficient to parameterize those methods
that deal with transitions.

Removing this parameter simplifies code: It avoids the need to
carry around the additional type parameter LETTER when it is not
needed, or using wildcards.