fix implementation of successor sleep map computation
The old implementation was incorrect: It looked up the so-far-computed
price of the letter on the transition (line 142), when the intention
was likely to take the price of the current letter in the old sleep map.

The new implementation is better documented and hopefully easier to
understand.

Further, it should improve performance (though perhaps only slightly)
by avoiding unnecessary recomputations and limiting the maximum relation
to check in advance rather than retroactively.