preliminary hackish support for dead ends with abstract and multi-POR
Add bookkeeping mechanism to determine down to which
abstraction level a state will be a dead end.

This is necessary when dealing with abstract independence,
either alone or in multi-reduction.

However, currently the implementation massively deteriorates
performance: A large percentage of overall time is spent on
bookkeeping, and the remaining time also significantly increases.
The reasons are not yet entirely known, but may be related to the
fact that bookkeeping happens at the outermost layer of independence
relations.