fix determination of loop locations
Previously, the loop heads of loops immediately following
a procedure entry were not marked as loop locations. This
change fixes this, by pulling the code out of the method
getLocNodeForLabel into the caller method processLabel,
where it is also executed in case the mCurrent is already
a location (and not a statement).

getLocNodeForLabel is also called from processGotoStatement,
but it should be enough to mark the node as loop location once
(when the label is processed) rather than repeat this for every
goto statement to this label.

The commit also fixes the debug log message.