Try out more liberal check for accepting transitions
The previous behaviour is sometimes overly cautious,
and results in the first transition of a composition
being preserved in many cases where this is clearly
not necessary.

The liberal check seems to fix this, but currently
causes an additional unsoundness, which is detected
by the test nonexec-2token-input-bug.ats.

However, this unsound behaviour appears to be more related
to the discarding of transitions if they are involved in a
non-executable composition pair. It is neither clear to me
why this is allowed, nor that the discarding in combination
with the strict check cover all cases.

Hence perhaps the solution is to change the discarding behaviour,
rather than enforce the strict check for accepting transitions.