fix bug with duplicated IcfgEdges
Due to incorrect calls between overloads of
IcfgEdgeBuilder#constructInternalTransition,
the same edge was added as outgoing edge of
a location multiple times.

As a result, removal of such edges did not work as
expected. Therefore, the ICFG contained references
to incomplete edges (with source and target NULL).
This lead to NullPointerExceptions.

This commit fixes IcfgEdgeBuilder and additionally
adds two small asserts to detect similar bugs more
easily in the future.