refactor SequenceRule
The code should be more readable now.

Also, instead of collecting a bunch of changes to the net
and then applying them, we now apply them directly, with
a corresponding update of the coenabled relation and the
retromorphism. Therefore there is no more need to ensure
that two compositions don't overlap. Thus, fewer iterations
of the sequence rule should be needed to reach a fixpoint
(ideally just one iteration).