Don't evaluate string literals passed to printf
Since we don't expect string literals to have any side effects
relevant to any property, and they are not used since we do not
model the effect of printf, there is no need to evaluate them.

This avoids having overapproximations of large literals which
needlessly prevents us from reporting errors.