fix bug in C backtranslation: incorrect variable types (and possibly names)
If two C variables are translated to Boogie variables with the same name
(this typically happens if the C variables have the same name), then the
translation only records the type of one of the variables.

The backtranslation thus may retrieve the wrong type. In cases where this
incorrect type was not a primitive (but e.g. a pointer), this resulted in
crashes in TypeSizes::extractIntegerValue.