improve ViolationWitnessGenerator::reduceProgramExecution
Previously, when encountering a sequence of AtomicTraceElements with
the same line number, only the first element was used for the witness.
Thus, for e.g. a call "plus(A[i], x)", only the element for "A[i]" was
used. This lead to witnesses with (among possibly other issues) broken
call stacks (later on, we return from "plus" but the witness does not
contain any call to "plus").

Instead, the solution implemented in this commit combines subsequent
trace elements with the same line number into one element:

- We use "step" and "element" of the last element in the sequence,
  as this should be the outermost expression/statement in the line
  (in the example above, the entire call to "plus").
- We combine the StepInfo, such that if any element is a call or a
  return, the entire combined element is also marked as such.
- Similarly for fork and join.

It would not be sufficient to only take the last element, as in this
case e.g. returns are lost if the calls appear in a larger statement,
say "x = plus(A[i], x);".