fix backtranslation of old values for pointer variables
This previously lead to crashes with "AssertionError: Expected only
CPrimitive but got ..." because the old values for pointer base and
offsets were treated as ints, but had an unexpected type (CPointer
rather than a CPrimitive).

This commit extends the special handling for (current values of)
pointers to their old values.

Note that the "AssertionError: Expected only CPrimitive but got ..."
still occurs in some cases, because the program state only contains
a value for a pointer's base (or only a value for the offset). This
may occur both for old and current values of pointers.