Use eliminateLight instead of eliminateCompat
The calls to (the old method) eliminateCompat were made with
argument ONLY_DER. This is quite similar to eliminateLight but
includes the old elimination of `div` and `mod`. The elimination of
`div` and `mod` can be costly and might cause nontermination in
rare cases.
I think the intension of the caller is to have an inexpensive
methods that removes quantifiers that are easy to eliminate. The
currently best method for this purpose is eliminateLight.