add an IChcScript to invoke eldarica via command line
The current EldaricaChcScript does not work well, we are using the
wrong entry point to eldarica. Until we have figured out the right
one, invoking eldarica via command line will seemingly give better
results.

This is a temporary solution. We should ideally find the right entry
point, or at least some other way that is not this platform-dependent.