Do not backtranslate memory related invariants
Currently the backtranslation of memory related invariants (like equality of pointers or dereference values) is not stable yet and might produce wrong invariants. Therefore we want to exclude this for this year's SV-COMP.
This commit ensures that we still have TemporaryPointerExpressions in the counterexamples (should only occur in the log), because it might be useful for understanding the incorrectness of a program.