refactor CfgBuilder LBE to fix bug
During CFG construction, we mark the start and end of atomic blocks.
Subsequently, the LBE tries to fuse all edges within an atomic block
to achieve the required atomicity.

Previously, the start/end markers for atomic blocks were attached to
locations of the CFG. This is however problematic, as a location may
e.g. have multiple outgoing edges, where only some of them belong to
an atomic block. As a consequence, we sometimes believed that a loop
was part of an atomic block and threw an exception (because loops in
atomic blocks are unsupported), when actually only the loop body was
atomic. This also occurred in several SV-COMP benchmarks.

With this refactor, we instead attach the markers to the CFG _edges_
themselves rather than to locations. This allows us to unambiguously
detect if there is illegal control flow in an atomic block.

As a side effect, we no longer support some exotic usages of SV-COMP
__VERIFIER_atomic_begin/_end which are in contradiction of the rules
anyway. E.g. a call to __VERIFIER_atomic_end that can not be matched
with a call to __VERIFIER_atomic_begin is no longer supported.