PetriFloydHoare: replace incorrect computation of reachable markings
The previous computation missed markings not represented by **local**
configurations of events in the Petri net.

The new computation is very similar in spirit to the already implemented
algorithm for the "iterativeCosets" option, but is much more optimized.
Most importantly, it uses a fixed ordering of the set of conditions to
avoid generating the same cut several times. It also uses memory-efficient
data structures like ImmutableList, and benefits from smaller optimizations.