NVP generation is now done without complete power set generation.
Strongly connected location sets are generated using the Tarjan algorithm, and only their power sets are checked (starting from the largest, so that subsets of detected NVPs can be skipped).
This can still be slow (O(2^{#locations})), but I think this is unavoidable, since we need to find the relevant set of states somehow.