fix block encoding bug discovered by contract computation
Every procedure that exists in the IIcfg should have an exit location,
even if that location is unreachable.

This was not the case for the IIcfgs produced by the BlockEncodingV2 plugin.
This commit fixes this bug, and introduces an assert statement after block
encoding has finished to check if the produced IIcfg satisfies this invariant.

Moreover, we add two sanity checks regarding the computation of procedure contracts,
where the problem was discovered (the location of the ProcedureContractResult was null),
such that similar issues are caught earlier in the future.