remove unused method IAbstractPredicate::getProcedures
It is not clearly defined (let alone documented) what the procedures
of a predicate are. Since this method is not used (except to provide
procedures to other predicates), this commit removes it.

Also remove procedure collection from TermVarsProc, as this was only
used to provide procedures for IPredicates. Consequently, rename the
class TermVarsFuns to reflect its current functionality.