fix ILocation of labels
The DebugIdentifier for a mention of a label at a goto statement and
the mention of a label at the label's definition are never reference-
equal, but they are equal.

Fixing this condition allows the correct ILocation to be attached to
the label, such that an invariant for this label is assigned to the
correct location in witnesses / results. Previously, the common AST
parent node of label and goto was used, leading to rejected invariants.