Fix Floyd/Hoare validity check for Icfgs: actually check something
Previously, the check did not actually check anything, because the
successor edges were calculated incorrectly and, in particular for
initial states, were always empty.

We also need to filter summary edges for implemented procedures as
they are not expected to be inductive. This was previously done in
HoareAnnotationChecker [1] previous to PR #671.

[1] https://github.com/ultimate-pa/ultimate/blob/19fd058c3d438585a4f478206c978f85e13399ca/trunk/source/TraceAbstraction/src/de/uni_freiburg/informatik/ultimate/plugins/generator/traceabstraction/predicates/HoareAnnotationChecker.java#L141-L143