add support for encoding 'modifies' clauses in 'ensures'
As our witnesses do not support 'assigns' or 'modifies' clauses
in contracts, we encode non-modified variables by adding conjuncts
of the form 'x == \old(x)' to the 'ensures' clause.

This is a bit of a workaround. Ideally, this should be added only
in the backtranslation from Boogie to ACSL. However, this is not
easily possible in the current architecture.

# Conflicts:
#	trunk/source/Library-Proofs/src/de/uni_freiburg/informatik/ultimate/lib/proofs/floydhoare/FloydHoareUtils.java