add new interface ISymbolicIndependenceRelation
Implementations of this interface generate sufficient conditions for
commutativity of two given letters wrt some commutativity relation.

IIndependenceRelation now includes a method to retrieve a corresponding
ISymbolicIndependenceRelation. Implementations for the various types of
independence relations are included in this commit.

Incorporating commutativity condition generation along the stack of
independence relations ensures that the generated conditions are indeed
suitable for whatever stack of relations is used, while maintaining the
flexibility that comes with the ability to build different stacks.

For instance, with this architecture, one can easily decide to apply an
abstraction before checking independence (ie, to use an instance of the
type IndependenceRelationWithAbstraction) without worrying how to adapt
commutativity condition generation such that the conditions reflect the
abstract independence.