Fix translation of atomic functions (#685)
Previously we translated C-expressions possibly twice due to the optimization for pointer reads/writes and made everything atomic (to handle stdatomic functions).
Now, we make sure that the expression is only translated once and we separate the ExpressionResult of the original expressions and the corresponding read/write, s.t. we can control the atomicity for stdatomic functions properly.
Therefore we added separate method based on LocalLValue and HeapLValue.