rewrite IRule interface and properly handle independence
To properly handle independence of actions involving local variables,
we must know the thread instances involved in the action. Hence, switch
from an independence defined on IRule to an independence defined on a
new class RuleInstance and use a DisjointLocalsIndependenceRelation.

To implement reduction using this new notion of independence, refactor
once again the IRule interface. Significantly simplify sleep set reduction
in the process.