move predicate transferral out of TransferrerWithVariableCache
This is needed to support sharing an independence cache across CEGAR
loops (b03bd7cf) while also using symbolic relations.

This also avoid the need for a hack introduced in 8c24d7a6 where a
TransferrerWithVariableCache was responsible for creating its own
predicate factory, but had to expose it as it was also needed by the
caller.