properly handle multiple static variables with the same name
When multiple static C variables in different functions have the same
name, we incorrectly identified them as multiple declarations of the
same variable and only kept one of the corresponding Boogie declarations
(nondeterministically choosing which one).

To fix this, we now remember the scope for static object declarations,
and only identify declarations with the same name in the same scope as
multiple declarations of the same variable.

This issue caused e.g. undeclared identifier errors in the SV-COMP task
c/memsafety-ext3/scopes4-1.c