Bugfix: List UnprovabilityReasons
Objects of Overapprox store a map from reason to location. While summarizing all overapproximations of a program execution or trace, we just merged the maps. Hence, we kept only one location for each reason. This commit replaced the Maps with HashRelations.

But there is probably an even bigger problem: if we merge Overapprox objects, we also lose information if there was an overapproximation at several locations for a similar reason. We probably also have to replace Maps with HashRelations in the Overapprox class.