enable commutativity condition synthesis in GemCutter default settings
For now, it is only enabled for unreach-call tasks, as we are using
abstract commutativity for the other properties, and the combination
is not yet supported.

In integer translation mode, we are using the default refinement
strategy for commutativity condition synthesis, i.e., SmtInterpol.
In bitvector mode, we use the same strategy as for infeasibility
proofs, i.e. FOX.