fix crash in backtranslation of lasso traces
Previously, when finding a non-terminating lasso execution, we back-
translated stem and loop separately. This lead to crashes, as the call
stack of the loop may not be balanced. E.g. it may contain a return
before the first call, because the matching call is part of the stem
resp. the previous lasso iteration.

To solve this, we now backtranslate stem and loop together. The default
implementation is simply to translate them separately, but for those
backtranslators that had issues with this in SV-COMP benchmarks (namely,
inlining and C translation), we add special handling.

It would still be good in the future to analyse this issue more closely
and see if other backtranslators should also have special handling (e.g.
if no inlining is used). Also, there may still be some assert statements
in the mentioned backtranslators that fail. I'm ignoring them due to lack
of time and because asserts are disabled at SV-COMP.