remove unnecessary parameters from EmpireValidityCheck
EmpireValidityCheck should be independent of how an empire is constructed,
and should not rely on a "refined net" or a notion of "assertion places".
In fact, as the implementation only used transitions of the "original net"
(i.e., the net for which an Empire is given), these parameters already seem
unnecessary, and can be removed.