Use SV-COMP settings for Referee in web interface
Previously, a custom settings with bitvectors was used.
There is however an issue regarding undefined behavior in Referee (not in the program itself, but along the paths checked by Referee).
To reduce the impact of this issue, integer translation is used instead to bitvector translation, where there is at least no problem with overflows.

Additionally, the SV-COMP settings are better maintained than specific settings for the web interface.