ExpressionBacktranslation for Inliner: fix handling of renamed parameters
Parameters in declaration and implementation of a procedure may have
different names. The code attempted to check for this case and fail,
but this was prevented by a typo (comparing a name to itself, not the
other name).

This commit implements the translation by committing to use the name
in the implementation. As commented, this may have to change in case
we ever support backtranslation of procedure contracts over inlining.