remove broken ACSL check
This check is broken, in the sense that it is always fals and does not achieve
the intended effect: an ACSL Contract instance never implements the interface
IASTFunctionDefinition. The intent seems to have been to check whether the
annotated element of the C AST is an IASTFunctionDefinition.

As contracts are now supported, the check is also obsolete and can be removed.