division by zero: avoid trivial ("assert true") statements
In particular, such an "assert true" statement was generated for the
following declaration in some C header (and thus in many pre-processed
SV-COMP benchmark tasks):

    __cpu_mask __bits[1024 / (8 * sizeof (__cpu_mask))];

As a result, CHandler crashed due to not supporting side-effects
(e.g. statements) as part of a DeclarationResult.

A similar filtering of trivial asserts has previously already been
implemented for overflow checks.