Remove Jacoco from Sonar profile
We run Jacoco in the coverage profile anyways that is a previous step in the nightly build.
Therefore, it should not be necessary to run it again in the Sonar profile, as in the nightly the reports are still there (since we don't use mvn clean).
Otherwise we run it twice on the dev branch and therefore, the Jacoco results are also shown twice in Jenkins.