Reduce number of nodes in ICFG
Merge nodes constructed for labels directly with successor node even if
the successor node is a loop head. As Dominik noted today this is sound
even if we compute invariants: After a backtranslation of invariants,
the invariant at the label might be unneccesarily weak but it is not
weaker as the successive invariant (the very same invariant of the loop
head).