Reduce number of nodes in ICFG
If we have a `while(true)` loop and want to omit the `assume true` for
entering the loop, we merge the loop head with the first location of the
loop body even if this is stems from a label. Rationale: the merge is
sound even for the computation of invariants because there is no
execution the passes the loop head but does not enter the loop.