LargeBlockEncoding: ensure iteration over locations from start to end
IcfgBuilder and RcfgBuilder construct their respective CFGs in different
orders, resulting in different iteration orders for LargeBlockEncoding.
We suspect this is the main cause for the different behaviour of LBE on
CFGs constructed by these two plugins.

In order to perform this iteration, the CFG's initial nodes must be set
before LargeBlockEncoding is called.