Workaround: Duplicate ConstantTermNormalizer
Even though this class is present in SMTInterpol, it is only used by Ultimate and it was
modified to fix a bug (see ddd1310 and 6d19ff8). Therefore, it was copied to
Library-SmtLibUtils and SMTSolverBridge (it has to be duplicated, as there is no matching
common import of those and they cannot imported by each other because of cyclic
dependenices). But the long term solution should be to fix ConstantTermNormalizer in
SMTInterpol, i.e., https://github.com/ultimate-pa/smtinterpol/pull/150 should be
revisited and merged.