Refactor handling of standard functions and predefined types (#714)
Instead of handling all the standard functions from any library in
StandardFunctionHandler (which was a huge class for this reason), there is now an
interface ILibraryModel with implementations for different libraries (e.g., StdlibLibrary
model for stdlb.h, AtomicLibraryModel for stdatomic.h), where only the corresponding
functions are handled. Additionally, the implementations of ILibraryModel also contain
the types that are defined in the corresponding header, which was done in TypeHandler
before (similarly for different libraries).