BuchiAutomizer: fix handling of statistics for multiple CEGAR loops
In case of concurrent programs, we may start multiple CEGAR loops when
the initial number of thread instances was insufficient. However, this
lead to assertion errors and crashes as the same statistics object was
re-used and could not handle the duplicated data.

Now we proceed similarly to TraceAbstraction and create new statistics
objects for each CEGAR loop.