Add ACSLProblemNode to crash only if ACSL should be checked
Before, we always canceled the toolchain, if an error during parsing of ACSL occured (see CommentParser). However,
since ACSL checks can now be disabled, we should not crash if they are disabled. Instead, we return an ACSLProblemNode
if parsing fails and we only throw an exception when handling such anACSLProblemNode in ACSLHandler, which is only
called if ACSL should be checked.