Remove struebli from target file
The server is currently not working, therefore a build in Eclipse fails after reloading the target platform.
Since currently no dependencies from struebli were used, it was safe to remove it.
In case we need it in the future (e.g., to use self built dependencies), we can just add it again.