Refactor to use Boogie AST Unit as translation result
Replaces WrapperNode with de.uni_freiburg.informatik.ultimate.boogie.ast.Unit as the result type for the LlvmirToBoogie translation process. Updates observer and listener classes to reflect this change and adds construction of an empty Unit in the listener. Also updates MANIFEST.MF to import the required Boogie AST package.