Owicki-Gries tests: export and parse possible interferences
Modify OwickiGriesTestDumper to export possible interferences relation
(aka "co-marked" relation) in comments of the generated .ats file.

Modify OwickiGriesTestSuite to parse this relation. If the relation is not
specified, it is still computed from an unfolding. However, we must use an
unfolding of the program, rather than the difference. This was implemented
incorrectly, and thus might have missed non-interference violations.

Modify proof producers to take possible interferences as constructor input
rather than attempting to compute them from their other inputs.

For test suites that no longer require an unfolding of the difference net,
avoid computing such an unfolding.