Avoid unnecessary loop in memmove / memcpy (fixes #731)
This loop is only necessary, if pointers on the heap are needed.
However, this check was done inside an auxiliary method (constructMemcpyOrMemmovePointerLoopAssignment) that just created a simple counting loop otherwise.
Instead, we omit this loop entirely if pointers on the heap are not needed, by moving this check of this auxiliary method.
Thanks to @jankoerner for the observation!