CACSL2Boogie: AllocInit Counter to memory model
The the step size of the base address counter incremention depends now
on the exact implementation of the memory model.
With the 1D-addressing mode, the step size always increase with the
total size of allocated object.
With the 2D-addressing mode, it increases with a step size of one per
allocated object.