Separate build of Ultimate products from creating ZIP archives
This change introduces a new makeBuild.sh script that builds all
Ultimate products and packages the products with the required
configurations (toolchain and settings files), additional programs
and files like solvers or licenses. This script is then used in the
existing makeFresh.sh script for the common build steps to avoid
duplicate script code. The functionality of the makeFresh.sh script
remains unchanged.