partial improvement of TypeHandler::isCompatibleType (#693)
Our implementation of this method seems rather incomplete. This commit
is not yet a full solution to this incompleteness. That will require a
thorough study of the C standard.

This change fixes the errors occurring for the examples in issue #693
(initializers for nested structs). The issue was that the type of the
initializer expression was not recognised as compatible with the type
of the struct field. As a result, the constructIndexToInitInfo method
in the InitializationHandler class entered the wrong branch in l1275,
thus treating the initializer expression as belonging to the field of
the nested struct rather than the enclosing struct.