O/G proof producers: pass on-demand implementation of difference
None of the current implementations depend on having an explicit
representation, which may be costly to create.
If any implementation does require this, it should compute the
explicit representation itself.

This will also speed up our test suite, as it allows NaiveOG to
no longer require a finite prefix of the unfolding.