Bugfix MemorySlicer: Detect direct memory accesses in call statements
#703 replaced read~unchecked by simple array accesses to the #memory array.
However, while this change was implemented in MemorySlicer, the handling within calls was forgotten.
Therefore, we crashed while building the CFG with a strange exception.
Now, we just call processExpressions when constructing new calls.
For now, this just throws an exception (that is then catched and just omits the MemorySlicer), but this case might be handled properly in the future.