Fix location of target waypoint in YAML witness
Previously, we always used the exact expression where the violation of property occurred.
However, according to the documentation [1] it should be only a full expression or statement.
The implementation was initially only used for the unreach-call property, where this did not make any difference,
as the violation was always a call to reach_error() (which is already a full expression).

[1] https://gitlab.com/sosy-lab/benchmarking/sv-witnesses/-/blob/main/user-guide/Witness-Format.md#target