Boogie2ACSL: replace "isNegated" with "permissible approximation"
When backtranslating invariants, we want to omit untranslatable conjuncts
of the invariant, but keep the translatable conjuncts. To support this,
the translation previously kept a boolean "isNegated": under a negation,
we cannot omit conjuncts, but we can instead omit disjuncts.

However, there are many contexts in which we can omit neither conjuncts
nor disjuncts (essentially, non-monotone contexts). To better reflect this,
we introduce three separate modes: OVERAPPROXIMATION (omit conjuncts),
UNDERAPPROXIMATION (omit disjuncts), and PRECISE (no omissions).